Refactor Manager and Worker interface
